bugs fixed

+++ controllers/media.js
@@ -3,6 +3,9 @@ var mv = require('mv');
 var uuid = require('node-uuid');
 var fs = require('fs');
 var path = require('path');
+var ffmpeg = require('liquid-ffmpeg');
+var easyimg = require('easyimage');
+var async = require("async");
 
 // Insert a new Slider record
 exports.upload = function(req,res){
@@ -16,12 +19,9 @@ exports.upload = function(req,res){
     }else{
 
     	// create media record for database
-
     	var newMedia = new Media();
-
     	newMedia.title = req.body.title;
-    	newMedia.size = req.files.file.size;
- 
+    	newMedia.size = req.files.file.size; 
     	newMedia.target = req.body.target;
 
     	// determine the file type by file extension name
@@ -43,7 +43,7 @@ exports.upload = function(req,res){
     		newMedia.type = 'Other';
     	}
     	
-
+    	// combine the file path
     	var subPath = '/docs/';
     	subPath += newMedia.type + '/';
     	subPath += newMedia.target + '/';
@@ -51,51 +51,142 @@ exports.upload = function(req,res){
 		var newFileName = uuid.v4() + '.' +req.files.file.extension;
     	// new file location in server
    		var newPath =  'public/' + subPath + newFileName;
-
-
-   		// move file the new folder
-    	mv(req.files.file.path, newPath,{mkdirp: true},function(err){
-    		console.log('error form moving file');
-    		console.log(err);
-    	});
-
+   		var thumbnail_path = 'public/docs/thumbnails/'+ newFileName;
     	// doucment record path for http access
     	newMedia.path = subPath +  newFileName;
+    	newMedia.thumbnail = '/docs/thumbnails/' +  newFileName;
 
 
 
-    	// create video of image thumbnail
-    	if(newMedia.type == 'Video'){
-    		var ffmpeg = require('fluent-ffmpeg');
-				var proc = new ffmpeg('/path/to/your_movie.avi')
-				.takeScreenshots({
-				  count: 1,
-				  timemarks: [ '600' ] // number of seconds
-				}, '/path/to/thumbnail/folder', function(err) {
-				console.log('screenshots were saved')
-			});
+    	async.series([
 
-    	}else if (newMedia.type == 'Image'){
+		    function(next){
+		    	//---------------------------------------
+		    	// create media thumbnail
+		    	//---------------------------------------
 
-    	}else{
+		    	if(newMedia.type == 'Video'){
+		    		//create video of image thumbnail	
+		    		console.log('start thunbnail process');
+		    		console.log(__dirname + '/' + req.files.file.path);
 
-    	}
+		    		var source_path = '/sites/nodejs/ecs/' + req.files.file.path;
+		    		console.log(source_path);
+					
 
-    	// save db record
-	    newMedia.save(function(err,result){
-			if(err){
+					var proc = new ffmpeg({ source: req.files.file.path })
+					  .withSize('150x100')
+					  .takeScreenshots({
+					      count: 1,
+					      timemarks: [ '50%' ]
+					    }, 'uploads/' , function(err, filenames) {
+
+					    	if(err){
+					    		console.log(err);
+					    	}else{
+					    		console.log(filenames);
+					      		console.log('screenshots were saved');
+					    	}
+					   
+					   		next();
+					  });
+
+		    	}else if (newMedia.type == 'Image'){
+
+		    		// create image thumbnail
+
+		    		
+
+					// easyimg.rescrop({
+					//      src: req.files.file.path, dst: thumbnail_path,
+					//      width:500, height:500,
+					//      cropwidth:128, cropheight:128,
+					//      x:0, y:0
+					//   }).then(
+					//   function(image) {
+					//      newMedia.thumbnail = thumbnail_path;
+					//      console.log('Resized and cropped: ' + image.width + ' x ' + image.height);
+					//      next();
+					//   },
+					//   function (err) {
+					//     console.log(err);
+					//     next();
+					//   }
+					// );
+
+					easyimg.info(req.files.file.path).then(
+						function(file) {
+
+							easyimg.resize({
+							     src: req.files.file.path, dst: thumbnail_path,
+							     width:250 * file.width / file.height , height:250
+							  }).then(
+							  function(image) {
+							     //newMedia.thumbnail = thumbnail_path;
+							     console.log('Resized and cropped: ' + image.width + ' x ' + image.height);
+							     next();
+							  },
+							  function (err) {
+							    console.log(err);
+							    next();
+							  }
+							);
+
+						}, function (err) {
+
+							console.log(err);
+							next();
+						}
+					);
+
+
+
+
+
+		    	}else{
+
+		    	}
+        
+
+		    },
+
+
+		    function(next){
+
+		        //---------------------------------------
+		        //  move file the new folder
+		        //---------------------------------------
+
+		    	mv(req.files.file.path, newPath,{mkdirp: true},function(err){
+		    		console.log('error form moving file');
+		    		console.log(err);
+		    		next();
+		    	});      
+		       
+		    }
+		],
+		// optional callback 
+		function(err, results){
+
+
+
+	    	// save db record
+		    newMedia.save(function(err,result){
+				if(err){
+					res.json({
+						type:false,
+						data:"Error occured: " +err
+						});
+				}
 				res.json({
-					type:false,
-					data:"Error occured: " +err
-					});
-			}
-			res.json({
-				type:true,
-				data:result
-			});
+					type:true,
+					data:result
+				});
+			});	
+			
+
 		});
 
-    	//res.status(204).end();	
     }
 
 }
@@ -205,6 +296,8 @@ exports.delete = function(req,res){
 
 
 	Media.find({_id:remove_id}, function(err, obj){
+
+		// return error message if error 
 		if(err) {
 			res.json({
 				status: 'fail',
@@ -218,12 +311,12 @@ exports.delete = function(req,res){
 
 		file_path = file_path.replace("controllers", "public"); 
 
-
-		fs.unlinkSync(file_path);
-		// fs.unlink(file_path, function(err){
-		// 	console.log(err);			
-		// });
-
+		// remove file if exist
+		if (fs.existsSync(file_path)) {
+		    fs.unlinkSync(file_path);
+		}
+		
+	
 
 		Media.find({ _id:remove_id }).remove(function(err,result){
